Made two performance optimizations, the second of which makes the former less important. Also added a forgotten documentation file for "alphaorder".
1. Noticed that GEN is not constant-time, which is bad. On closer
inspection the problem was that the abstraction that's used to
instantiate the proforma is created separately from the abstracted
input theorem, so the pointer-EQ subterms are hidden under an
abstraction. As well as fixing that, made it more efficient in two
other ways: used EQ_MP instead of PROVE_HYP, and added partial
evaluation based on the first argument.

2. Made a small but significant change to the kernel, based on an
observation by Ondrej Kuncar, optimizing the pointer-EQ shortcut in
"orda" so that it does not require an empty environment but merely one
full of identical pairs. This can substantially speed up many derived
rules by allowing efficient use of proformas even inside bound
variables.



git-svn-id: http://hol-light.googlecode.com/svn/trunk@83 1eddc5a6-173b-11df-bff1-5f4ea15f2b7c