Fixed a LET_TAC bug pointed out by Kukhak Ky: failures were occurring if the same variable was used on the left and right of a pattern (except if they are a trivial let, i.e. corresponding lefts and rights exactly equal), e.g.
  g `v IN V DIFF (let k,ul = (P:num#(real^3)list->num#(real^3)list) (k,ul) in
     (Q:num#(real^3)list->real^3->bool) (k,ul)) ==> F`;;
  e LET_TAC;;

Now the abbreviating variables are renamed as necessary to avoid
clashes with free variables of the goal or each other (some of this
would previously happen later in the ABBREV_TAC step anyway in some
cases, but not this one).

Looking at such corner cases made me add a parsing warning if there
are multiple uses of the same variable in different arms of a
simultaneous let binding, e.g. `let x = 1 and x = 2 in x = 42`, in
which case the first one is hidden.



git-svn-id: http://hol-light.googlecode.com/svn/trunk@95 1eddc5a6-173b-11df-bff1-5f4ea15f2b7c